Extracted request creation logic
