+++ src/log4stash/ElasticClient/IRequestFactory.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace log4stash.ElasticClient
+{
+    public interface IRequestFactory
+    {
+        RequestDetails PrepareRequest(IEnumerable<InnerBulkOperation> bulk);
+    }
+}
