+++ src/log4stash/ElasticClient/RequestDetails.cs
@@ -1,0 +1,16 @@
+﻿using RestSharp;
+
+namespace log4stash
+{
+    public class RequestDetails
+    {
+        public RequestDetails(RestRequest restRequest, string content)
+        {
+            RestRequest = restRequest;
+            Content = content;
+        }
+
+        public RestRequest RestRequest { get; private set; }
+        public string Content { get; private set; }
+    }
+}
